style: move services into separated file
